Exception handling for users, category, subcategory tables

+++ Capstone/BackEnd/assessmentPortal/src/main/java/com/capstone/assessmentPortal/service/serviceImplementation/CategoryServiceImplementation.java
@@ -1,11 +1,15 @@
 package com.capstone.assessmentPortal.service.serviceImplementation;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.capstone.assessmentPortal.exception.AlreadyExistsException;
+import com.capstone.assessmentPortal.exception.EmptyListException;
+import com.capstone.assessmentPortal.exception.InputEmptyException;
 import com.capstone.assessmentPortal.model.Category;
 import com.capstone.assessmentPortal.repository.CategoryRepo;
 import com.capstone.assessmentPortal.service.CategoryService;
@@ -17,33 +21,54 @@ public class CategoryServiceImplementation implements CategoryService{
 
   @Override
   public Category addCategory(Category category) {
-    return categoryRepo.save(category);
+	Optional<Category> existingCategory = categoryRepo.getCategoryByName(category.getCategoryName());
+	if(existingCategory.isPresent()) {
+		throw new AlreadyExistsException();
+	}else {
+	  if(category.getCategoryName() == "") {
+		throw new InputEmptyException();
+	  }
+      return categoryRepo.save(category);
+	}
   }
 
   @Override
   public List<Category> getAllCategories() {
-    return categoryRepo.findAll();
+	List<Category> listOfCategories = categoryRepo.findAll();
+	if(listOfCategories.size() == 0) {
+	  throw new EmptyListException();
+	}else {
+	  return listOfCategories;
+	}
   }
 
   @Override
-  public Optional<Category> getCategoryById(Long categoryId) {
-    return categoryRepo.findById(categoryId);
+  public Category getCategoryById(Long categoryId) {
+    return categoryRepo.findById(categoryId).orElseThrow(()->new NoSuchElementException("Cannot find category with id: "+categoryId));
   }
 
   @Override
   public void deleteCategory(Long categoryId) {
-    categoryRepo.deleteById(categoryId);
+	Category existingCategory = categoryRepo.findById(categoryId).orElse(null);
+	if(existingCategory == null) {
+	  throw new NoSuchElementException();
+	}else {
+      categoryRepo.deleteById(categoryId);
+	}
   }
 
   @Override
   public Category updateCategory(Long categoryId, Category category) {
-    Category existingCategory = categoryRepo.findById(categoryId).orElseThrow();
+    Category existingCategory = categoryRepo.findById(categoryId).orElse(null);
 	if(existingCategory!=null) {
 	  existingCategory.setCategoryName(category.getCategoryName());
 	  existingCategory.setCategoryDescription(category.getCategoryDescription());
+	  if(existingCategory.getCategoryName() == "") {
+	    throw new InputEmptyException();
+	  }
 	  return categoryRepo.save(existingCategory);
 	}else {
-	  return null;
+	  throw new NoSuchElementException();
 	}
   }
 }
